Extract numpad into reusable component, fix button sizing
- Add src/components/ui/numpad.tsx with aspect-square buttons and max-w-sm
- Use NumpadKey type for type-safe key handling
- Remove inline numpad code from TransferInputStep
